refactor: 캐시 철자 수정
