+++ src/components/features/Order/atoms/OrderCashReceiptInput.tsx
@@ -2,6 +2,6 @@ import type { InputProps } from '@chakra-ui/react';
 import { Input } from '@chakra-ui/react';
 import { forwardRef } from 'react';
 
-export const OrderCacheReceiptInput = forwardRef<HTMLInputElement, InputProps>((props, ref) => (
+export const OrderCashReceiptInput = forwardRef<HTMLInputElement, InputProps>((props, ref) => (
   <Input ref={ref} placeholder="(-없이) 숫자만 입력해주세요." {...props} />
 ));
