+++ src/components/features/Order/organisms/OrderReceipt.tsx
@@ -5,8 +5,8 @@ import { Button } from '@/components/common/Button';
 import { OrderSubTitle } from '@/components/features/Order/atoms/OrderSubTitle';
 import type { IOrderPrice } from '@/components/features/Order/molecules/OrderPrice';
 import { OrderPrice } from '@/components/features/Order/molecules/OrderPrice';
-import type { IOrderRequestCashReceipt } from '@/components/features/Order/molecules/OrderRequestCacheReceipt';
-import { OrderRequestCacheReceipt } from '@/components/features/Order/molecules/OrderRequestCacheReceipt';
+import type { IOrderRequestCashReceipt } from '@/components/features/Order/molecules/OrderRequestCashReceipt';
+import { OrderRequestCashReceipt } from '@/components/features/Order/molecules/OrderRequestCashReceipt';
 
 export interface IOrderReceipt extends IOrderPrice, IOrderRequestCashReceipt {}
 
@@ -21,7 +21,7 @@ export const OrderReceipt = memo(({ totalPrice, register }: IOrderReceipt) => (
     <VStack spacing="24px">
       <OrderSubTitle subtitle="결제 정보" textAlign="left" />
       <Divider borderColor="#ededed" />
-      <OrderRequestCacheReceipt register={register} />
+      <OrderRequestCashReceipt register={register} />
       <Divider borderColor="#ededed" />
       <OrderPrice totalPrice={totalPrice} />
       <Divider borderColor="#ededed" />
